ride_project/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
gunicorn==21.2.0
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
